Adicionar componente Dropdown e SelectColor
