+++ src/components/dropdown/styles.module.scss
@@ -1,0 +1,33 @@
+.dropdownMenuContent {
+  min-width: 320px;
+  background-color: white;
+  border-radius: 6px;
+  padding: 0.375rem;
+  box-shadow: 0px 10px 38px -10px rgba(22, 23, 24, 0.35), 0px 10px 20px -15px rgba(22, 23, 24, 0.2);
+  animation-duration: 400ms;
+  animation-timing-function: cubic-bezier(0.16, 1, 0.3, 1);
+  will-change: transform, opacity;
+
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 0.5rem;
+}
+
+.dropdownMenuArrow {
+  fill: var(--white);
+}
+
+.dropdownMenuContent[data-side='bottom'] {
+  animation-name: slideUpAndFade;
+}
+
+@keyframes slideUpAndFade {
+  from {
+    opacity: 0;
+    transform: translateY(2px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
