--- src/components/settings/styles.module.scss
@@ -1,81 +1,0 @@
-.dropdownMenuTrigger {
-  border-radius: 6px;
-  border: 0;
-  background-color: var(--white);
-  border: 1px solid var(--gray-300);
-  color: var(--gray-600);
-
-  height: 3.5rem;
-  width: 3.5rem;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: all 200ms;
-
-  &:hover {
-    color: var(--gray-900);
-  }
-}
-
-.dropdownMenuContent {
-  min-width: 320px;
-  background-color: white;
-  border-radius: 6px;
-  padding: 0.375rem;
-  box-shadow: 0px 10px 38px -10px rgba(22, 23, 24, 0.35), 0px 10px 20px -15px rgba(22, 23, 24, 0.2);
-  animation-duration: 400ms;
-  animation-timing-function: cubic-bezier(0.16, 1, 0.3, 1);
-  will-change: transform, opacity;
-
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 0.5rem;
-}
-
-.dropdownMenuContent[data-side='bottom'] {
-  animation-name: slideUpAndFade;
-}
-
-.dropdownMenuItem {
-  font-size: 1rem;
-  line-height: 1;
-  color: var(--gray-600);
-  border-radius: 4px;
-
-  display: flex;
-  align-items: center;
-  height: 2rem;
-  padding: 0 0.5rem;
-  position: relative;
-
-  user-select: none;
-  outline: none;
-}
-
-.itemColor {
-  width: 1rem;
-  height: 1rem;
-  border-radius: 50%;
-  margin-right: 0.5rem;
-}
-
-.dropdownMenuItem[data-highlighted] {
-  background-color: var(--gray-200);
-  color: var(--gray-900);
-}
-
-.dropdownMenuArrow {
-  fill: var(--white);
-}
-
-@keyframes slideUpAndFade {
-  from {
-    opacity: 0;
-    transform: translateY(2px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
