+++ src/pages/home/index.tsx
@@ -1,8 +1,8 @@
 import { Banner } from '../../components/banner'
 import { Container } from '../../components/container'
 import { Logo } from '../../components/logo'
-import { Settings } from '../../components/settings'
 import { AddTodoForm } from './add-todo-form'
+import { Settings } from './settings'
 import styles from './styles.module.scss'
 import { TodoList } from './todo-list'
