+++ src/pages/home/settings/select-color/styles.module.scss
@@ -1,0 +1,27 @@
+.menuItem {
+  font-size: 1rem;
+  line-height: 1;
+  color: var(--gray-600);
+  border-radius: 4px;
+
+  display: flex;
+  align-items: center;
+  height: 2rem;
+  padding: 0 0.5rem;
+  position: relative;
+
+  user-select: none;
+  outline: none;
+}
+
+.menuItem[data-highlighted] {
+  background-color: var(--gray-200);
+  color: var(--gray-900);
+}
+
+.menuItemColor {
+  width: 1rem;
+  height: 1rem;
+  border-radius: 50%;
+  margin-right: 0.5rem;
+}
